refactor: use entities instead of collisionstorage
