+++ src/plugin/collision/src/Collision.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "2DCollisions.hpp"
+#include "CollisionUtils2D.hpp"
+#include "CollisionUtils3D.hpp"
 #include "BoxCollider2D.hpp"
 #include "BoxCollider3D.hpp"
 #include "CollisionChecker.hpp"
-#include "CollisionStorage.hpp"
