+++ src/plugin/collision/src/component/BoxCollider3D.hpp
@@ -5,7 +5,7 @@
 namespace ES::Plugin::Collision::Component {
 
 /**
- * Basic collider used to check collisions through CollisionChecker3D system
+ * Basic collider used to check collisions through CollisionCheckerABAB system
  */
 struct BoxCollider3D {
     /**
