+++ src/plugin/collision/src/utils/CollisionUtils2D.hpp
